moved stray slice function
