+++ cxfx/op_openal_desktop.go
@@ -3,7 +3,6 @@
 package cxfx
 
 import (
-	"github.com/skycoin/cx/cx"
 	"github.com/skycoin/cx/cx/ast"
 	"github.com/skycoin/cx/cx/helper"
 	"golang.org/x/mobile/exp/audio/al"
@@ -103,7 +102,7 @@ func opAlGenBuffers(inputs []ast.CXValue, outputs []ast.CXValue) {
 	outputSliceOffset := ast.GetPointerOffset(int32(outputSlicePointer))
 	for _, b := range buffers { // REFACTOR append with copy ?
 		obj := helper.FromI32(int32(b))
-		outputSliceOffset = int32(cxcore.WriteToSlice(int(outputSliceOffset), obj))
+		outputSliceOffset = int32(ast.WriteToSlice(int(outputSliceOffset), obj))
 	}
     outputs[0].SetSlice(outputSliceOffset)
 }
@@ -122,7 +121,7 @@ func opAlGenSources(inputs []ast.CXValue, outputs []ast.CXValue) {
 	outputSliceOffset := ast.GetPointerOffset(int32(outputSlicePointer))
 	for _, s := range sources { // REFACTOR append with copy ?
 		obj := helper.FromI32(int32(s))
-		outputSliceOffset = int32(cxcore.WriteToSlice(int(outputSliceOffset), obj))
+		outputSliceOffset = int32(ast.WriteToSlice(int(outputSliceOffset), obj))
 	}
     outputs[0].SetSlice(outputSliceOffset)
 }
